recorder/requirements.txt
pynput>=1.7.6
pyautogui>=0.9.54
mss>=9.0.1
Pillow>=10.0.0
requests>=2.31.0
